Backend/requirements_fixed.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyJWT==2.11.0
PyMySQL==1.1.2
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.6
Flask-Mail==0.10.0
python-dotenv==1.0.1

# NLP & Similarity Detection
sentence-transformers==2.3.1
scikit-learn==1.4.0
numpy==1.26.3

# AI Content Detection
transformers==4.37.2
torch==2.2.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Report Generation
reportlab==4.1.0
Pillow==10.2.0